+++ src/utils.ts
@@ -1,3 +1,4 @@
+import { Request } from 'express';
 import { Dictionary } from 'lodash';
 import isEmpty = require('lodash.isempty');
 import omit = require('omit-deep');
@@ -5,4 +6,6 @@ import omit = require('omit-deep');
 const removeEmpty = (obj: Dictionary<any>): object =>
     omit(obj, Object.keys(obj).filter(key => obj[key] === undefined || isEmpty(obj[key])));
 
-export { removeEmpty };
+const matchPath = (pattern: RegExp) => (req: Request): boolean => req.originalUrl.match(pattern) !== null;
+
+export { removeEmpty, matchPath };